feat: CoreDataManager Save 구현 및 DTO 프로퍼티 타입 수정 #20

+++ Left7/Data/ResponseModel/CoreData/ProductDataObject+CoreDataProperties.swift
@@ -15,12 +15,12 @@ extension ProductDataObject {
         return NSFetchRequest<ProductDataObject>(entityName: "ProductDataObject")
     }
 
-    @NSManaged public var id: Int16
+    @NSManaged public var id: Int
     @NSManaged public var name: String?
     @NSManaged public var thumbnailPath: String?
     @NSManaged public var descriptionImagePath: String?
     @NSManaged public var descriptionSubject: String?
-    @NSManaged public var price: Int16
+    @NSManaged public var price: Int
     @NSManaged public var rate: Double
     @NSManaged public var isFavorite: Bool
     @NSManaged public var favoriteRegistrationTime: Date?
